+++ src/main/java/dev/excal1bur/appliedsmelting/core/ModCreativeTab.java
@@ -17,8 +17,15 @@ public final class ModCreativeTab {
                 .icon(() -> ModBlocks.ME_SMELTER_ITEM.get().getDefaultInstance())
                 .displayItems((parameters, output) -> {
                     output.accept(ModBlocks.ME_SMELTER_ITEM.get());
+                    output.accept(ModBlocks.ME_SMELTER_MK1_ITEM.get());
+                    output.accept(ModBlocks.ME_SMELTER_MK2_ITEM.get());
+                    output.accept(ModBlocks.ME_SMELTER_MK3_ITEM.get());
                     output.accept(ModBlocks.SMELTING_TERMINAL_ITEM.get());
                     output.accept(ModItems.FUEL_EFFICIENCY_CARD.get());
+                    output.accept(ModItems.SMELTER_UPGRADE_TEMPLATE.get());
+                    output.accept(ModItems.MK1_UPGRADE_KIT.get());
+                    output.accept(ModItems.MK2_UPGRADE_KIT.get());
+                    output.accept(ModItems.MK3_UPGRADE_KIT.get());
                 })
                 .build());
     }